[Silver II] Title: 최소 힙, Time: 16 ms, Memory: 2916 KB -BaekjoonHub

+++ 백준/Silver/1927. 최소 힙/최소 힙.cc
@@ -11,7 +11,7 @@ int main() {
     cin >> n;
     while (n--)
     {
-        long long int x;
+        int x;
         cin >> x;
         if (x == 0)
         {
